Extracts Reservation verification into a separate class
