+++ app/models/reservation.rb
@@ -31,17 +31,6 @@ class Reservation < ApplicationRecord
     Reservation.where(date: date)
   end
 
-  def self.verify(verification_code, user)
-    reservation = Reservation.find_by(verification_code: verification_code)
-    if reservation.nil?
-      return false
-    end
-
-    reservation.verified_by = user
-    reservation.save
-    reservation
-  end
-
   private
 
   def does_not_exceed_office_capacity
